+++ .eslintrc.js
@@ -1,0 +1,11 @@
+module.exports = {
+  extends: ['molindo/javascript'],
+  env: {
+    es6: true,
+    jest: true,
+    node: true
+  },
+  parserOptions: {
+    sourceType: 'module'
+  }
+};