+++ src/__tests__/Logger.test.js
@@ -1,11 +1,14 @@
+/* eslint-disable no-console */
+import {it, vi, expect, beforeEach, describe, afterEach} from 'vitest';
 import Logger from '../Logger';
 
 describe('Logger', () => {
-  const originalStdout = process.stdout.write;
-  const originalStderr = process.stderr.write;
+  const originalStdout = console._stdout.write;
+  const originalStderr = console._stderr.write;
+
   beforeEach(() => {
-    process.stdout.write = jest.fn();
-    process.stderr.write = jest.fn();
+    console._stdout.write = vi.fn();
+    console._stderr.write = vi.fn();
   });
 
   it('logs plain messages in development', () => {
@@ -21,20 +24,18 @@ describe('Logger', () => {
     logger.warn("don't forget to get it out in time");
     logger.error('pizza is burned!');
 
-    expect(process.stdout.write.mock.calls).toEqual([
+    expect(console._stdout.write.mock.calls).toEqual([
       ['TRACE: making a salami pizza\n'],
       ['DEBUG: adding salami\n'],
       ['INFO: putting it in the oven\n'],
       ["WARN: don't forget to get it out in time\n"]
     ]);
-    expect(process.stderr.write.mock.calls).toEqual([
+    expect(console._stderr.write.mock.calls).toEqual([
       ['ERROR: pizza is burned!\n']
     ]);
-
-    logger.destroy();
   });
 
-  it('only logs infos, warnings and errors in development', () => {
+  it('logs all but trace in development', () => {
     const logger = new Logger({service: 'pizza-shop', isProduction: false});
     logger.trace('making a salami pizza');
     logger.debug('adding salami');
@@ -42,10 +43,8 @@ describe('Logger', () => {
     logger.warn("don't forget to get it out in time");
     logger.error('pizza is burned!');
 
-    expect(process.stdout.write.mock.calls.length).toBe(2);
-    expect(process.stderr.write.mock.calls.length).toBe(1);
-
-    logger.destroy();
+    expect(console._stdout.write.mock.calls.length).toBe(3);
+    expect(console._stderr.write.mock.calls.length).toBe(1);
   });
 
   it('logs all levels in production', () => {
@@ -56,25 +55,21 @@ describe('Logger', () => {
     logger.warn("don't forget to get it out in time");
     logger.error('pizza is burned!');
 
-    expect(process.stdout.write.mock.calls.length).toBe(4);
-    expect(process.stderr.write.mock.calls.length).toBe(1);
-
-    logger.destroy();
+    expect(console._stdout.write.mock.calls.length).toBe(4);
+    expect(console._stderr.write.mock.calls.length).toBe(1);
   });
 
   it('logs json messages in production', () => {
     const logger = new Logger({service: 'pizza-shop', isProduction: true});
     logger.info('pizza is ready!');
 
-    const output = JSON.parse(process.stdout.write.mock.calls[0][0]);
+    const output = JSON.parse(console._stdout.write.mock.calls[0][0]);
     expect(typeof output['@timestamp']).toBe('string');
     expect(output['@timestamp']).toMatch(/\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}/);
     expect(output.level).toBe('INFO');
     expect(output.level_value).toBe(20000);
     expect(output.message).toBe('pizza is ready!');
     expect(output.service).toBe('pizza-shop');
-
-    logger.destroy();
   });
 
   it('overloads the log functions to provide a logger name', () => {
@@ -89,14 +84,12 @@ describe('Logger', () => {
       name: 'cook'
     });
 
-    const stdoutCalls = process.stdout.write.mock.calls.map(call =>
+    const stdoutCalls = console._stdout.write.mock.calls.map((call) =>
       JSON.parse(call[0])
     );
 
-    expect(stdoutCalls[0]['logger_name']).toBe('cook');
-    expect(stdoutCalls[1]['logger_name']).toBe('cook');
-
-    logger.destroy();
+    expect(stdoutCalls[0].logger_name).toBe('cook');
+    expect(stdoutCalls[1].logger_name).toBe('cook');
   });
 
   it('handles a request id', () => {
@@ -107,7 +100,7 @@ describe('Logger', () => {
       requestId: '123'
     });
 
-    const stdoutCalls = process.stdout.write.mock.calls.map(call =>
+    const stdoutCalls = console._stdout.write.mock.calls.map((call) =>
       JSON.parse(call[0])
     );
 
@@ -119,8 +112,6 @@ describe('Logger', () => {
         service: 'service'
       }
     ]);
-
-    logger.destroy();
   });
 
   it('throws when no service name is provided', () => {
@@ -128,7 +119,7 @@ describe('Logger', () => {
   });
 
   afterEach(() => {
-    process.stdout.write = originalStdout;
-    process.stderr.write = originalStderr;
+    console._stdout.write = originalStdout;
+    console._stderr.write = originalStderr;
   });
 });